doc-rag/requirements.txt
beautifulsoup4==4.12.3
langchain==0.3.14
langchain_chroma==0.2.0
langchain_community==0.3.14
langchain_core==0.3.29
langchain_openai==0.2.14
langchain_text_splitters==0.3.5
python-dotenv==1.0.1
unstructured==0.16.12
